Complete testing all the functions.

+++ Inferential.h
@@ -5,8 +5,8 @@
 #ifndef INFERENTIAL_H
 #define INFERENTIAL_H
 
-double Cov(int* valueX, int* valueY, int size );
-double Corr(int* valueX, int* valueY, int size );
-double *LinearRegression (int *Xarray, int *Yarray, int size);
+double Cov(double* valueX, double* valueY, int size );
+double Corr(double* valueX, double* valueY, int size );
+double *LinearRegression (double *Xarray, double *Yarray, int size);
 
 #endif //UNTITLED_INFERENTIAL_H
