+++ mean.cpp
@@ -1,10 +1,10 @@
-#include <iostream>;
+#include <iostream>
 #include "mean.h"
 
 using namespace std;
 
 
-double mean(int array[], int arraySize) {
+double mean(double array[], int arraySize) {
 
     double sum = 0.0,
             mean;
@@ -12,7 +12,6 @@ double mean(int array[], int arraySize) {
     for (int i = 0; i < arraySize; i++) {
         sum += array[i];
     }
-
     mean = sum / arraySize;
