+++ mode.h
@@ -5,6 +5,6 @@
 #ifndef COSC2082_GROUP_ASSIGNMENT_1_MODE_H
 #define COSC2082_GROUP_ASSIGNMENT_1_MODE_H
 
-int mode(int array[], int arrSize);
+double mode(double array[], int arrSize);
 
 #endif //COSC2082_GROUP_ASSIGNMENT_1_MODE_H
